Sqlalchemy- removing circular dependency
