+++ server/endpoint_detection.py
@@ -930,18 +930,12 @@ class EndpointManager:
             )
 
     def get_endpoint_id_from_path(self, endpoint_path, project_id):
-        conn_endpoints = psycopg2.connect(os.getenv("POSTGRES_SERVER"))
-        endpoints_cursor = conn_endpoints.cursor()
-
-        query = """
-        SELECT identifier FROM endpoints WHERE path = %s AND project_id = %s
-        """
-        endpoints_cursor.execute(query, (endpoint_path, project_id))
-        result = endpoints_cursor.fetchone()
-
-        conn_endpoints.close()
-
-        if result:
-            return result[0]
+        with SessionManager() as db:
+            endpoint = db.query(Endpoint).filter(
+                Endpoint.path == endpoint_path,
+                Endpoint.project_id == project_id
+            ).first()
+        if endpoint:
+            return endpoint
         else:
             return None
